+++ src/components/slideshow/index.ts
@@ -1,0 +1,3 @@
+export { Slideshow, type SlideshowProps } from "./slideshow";
+export { Frame, type FrameProps } from "./slideshow-frame";
+export { Element, type ElementProps } from "./slideshow-element";
