+++ src/components/slideshow/slideshow-element.tsx
@@ -1,59 +1,8 @@
 "use client";
-import {
-  ReactElement,
-  createContext,
-  useContext,
-  useState,
-  Children,
-  cloneElement,
-  JSX,
-  useEffect,
-  useRef,
-} from "react";
+import { useContext, JSX, useRef } from "react";
 import { motion } from "motion/react";
-
-interface SlideshowContextType {
-  animationDistance: number;
-  animationTime: number;
-  width: number;
-  height: number;
-}
-
-const SlideshowContext = createContext({
-  animationDistance: 20,
-  animationTime: 20,
-  width: 0,
-  height: 0,
-} as SlideshowContextType);
-
-interface FrameContextType {
-  isCurrent: boolean;
-  isKeyFrame: boolean;
-  paddingSize?: number;
-}
-
-const FrameContext = createContext({
-  isCurrent: false,
-  isKeyFrame: false,
-  paddingSize: 0,
-} as FrameContextType);
-
-export interface FrameProps {
-  children: ReactElement<ElementProps>[];
-  // These are set by React.cloneElement.
-  isCurrent?: boolean;
-  isKeyFrame?: boolean;
-  paddingSize?: number;
-}
-
-export interface SlideshowProps {
-  advanceTime?: number;
-  inBetweenTime?: number;
-  animationDistance?: number;
-  animationTime?: number;
-  previewFrame?: number | null;
-  children: ReactElement<FrameProps>[];
-}
+import { SlideshowContext } from "./slideshow";
+import { FrameContext } from "./slideshow-frame";
 
 export type StringAnimationDirection =
   | "up"
@@ -99,118 +48,6 @@ function createAnimationDirection(x: number, y: number): AnimationDirection {
 }
 
 export const customAnimationDirection = createAnimationDirection;
-
-export function Slideshow({
-  advanceTime = 7,
-  animationDistance = 20,
-  inBetweenTime = 0.2,
-  animationTime = 1,
-  previewFrame = null,
-  children,
-}: SlideshowProps) {
-  const [currentFrame, setCurrentFrame] = useState(0);
-  const [animateIn, setAnimateIn] = useState(true);
-  const timeouts = useRef<NodeJS.Timeout[]>([]); // Track timeouts
-  const containerRef = useRef<HTMLDivElement>(null); // FIXME: we need to wait until this is set before rendering things.
-  const [size, setSize] = useState<{ width: number; height: number }>({
-    width: -1,
-    height: -1,
-  });
-
-  useEffect(() => {
-    const updateSize = () => {
-      if (containerRef.current) {
-        const { width, height } = containerRef.current.getBoundingClientRect();
-        setSize({ width: width, height: height });
-      }
-    };
-
-    const resizeObserver = new ResizeObserver(updateSize);
-    if (containerRef.current) {
-      resizeObserver.observe(containerRef.current);
-    }
-
-    return () => resizeObserver.disconnect();
-  }, []);
-
-  useEffect(() => {
-    if (previewFrame == null) {
-      const scheduleFrame = () => {
-        // Timeout to start animating out
-        timeouts.current.push(
-          setTimeout(() => {
-            setAnimateIn(false);
-
-            // Timeout for in-between state
-            timeouts.current.push(
-              setTimeout(() => {
-                // Timeout to advance slide
-                timeouts.current.push(
-                  setTimeout(() => {
-                    setCurrentFrame(
-                      (prevFrame) => (prevFrame + 1) % Children.count(children)
-                    );
-                    setAnimateIn(true);
-
-                    // Schedule next cycle
-                    scheduleFrame();
-                  }, inBetweenTime * 1000)
-                );
-              }, animationTime * 1000)
-            );
-          }, advanceTime * 1000)
-        );
-      };
-
-      scheduleFrame();
-
-      return () => {
-        // Cleanup timeouts when component unmounts
-        timeouts.current.forEach(clearTimeout);
-        timeouts.current = [];
-      };
-    } else {
-      setCurrentFrame(previewFrame);
-    }
-  }, [advanceTime, animationTime, children, inBetweenTime, previewFrame]); // Run effect when `children` change
-
-  return (
-    <SlideshowContext
-      value={{
-        animationDistance,
-        animationTime,
-        width: size.width,
-        height: size.height,
-      }}
-    >
-      <div ref={containerRef} className="h-full w-full">
-        {Children.map(children, (child, index) =>
-          cloneElement(child, {
-            isCurrent: currentFrame === index && animateIn,
-            isKeyFrame: index === (previewFrame ?? 0),
-          })
-        )}
-      </div>
-    </SlideshowContext>
-  );
-}
-
-// Warning: isCurrent and isKeyFrame are set by React.cloneElement.
-export function Frame({
-  children,
-  isCurrent,
-  isKeyFrame,
-  paddingSize = 0,
-}: FrameProps) {
-  if (isCurrent == null || isKeyFrame == null) {
-    throw new Error("Frame must be a child of Slideshow.");
-  }
-  return (
-    <FrameContext value={{ isCurrent, isKeyFrame, paddingSize }}>
-      <div className="fixed h-full w-full">{children}</div>
-    </FrameContext>
-  );
-}
 
 function calculateRelativePosition(pos: number, slideshowLength: number) {
   return slideshowLength * (pos / 100);
